+++ src/app/manifestGeneration/ManifestVersions.ts
@@ -5,5 +5,6 @@
 export enum ManifestVersions {
     v13 = "v1.3",
     v14 = "v1.4",
+    v15 = "v1.5",
     devPreview = "devPreview"
 }