+++ src/app/templates/gulp.config.js
@@ -34,7 +34,11 @@ const config = {
         {
             version: "devPreview",
             schema: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json"
-        }
+        },
+        {
+            version: "1.5",
+            schema: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.5/MicrosoftTeams.schema.json"
+        }    
     ]
 };